fix: add missing types
