+++ package.json
@@ -12,7 +12,7 @@
   },
   "scripts": {
     "build": "tsc -p tsconfig.build.json",
-    "test": "multi-tape --node-arg='-r esbuild-register' test/*/test-*.ts"
+    "test": "node -r esbuild-register ./node_modules/.bin/multi-tape test/*/test-*.ts"
   },
   "main": "./build/index.js",
   "license": "MIT",
