+++ src/KnxLog.ts
@@ -1,34 +1,12 @@
 import util from 'util';
-import factory from 'log-driver';
+import factory, { Logger, LogLevel } from 'log-driver';
 
-export enum LogLevel {
-  Trace = 'trace',
-  Debug = 'debug',
-  Info = 'info',
-  Warn = 'warn',
-  Error = 'error',
-}
-
-export interface LogDriverOptions {
-  level: LogLevel;
-  format: (level: LogLevel, msg: string, ...args: any[]) => string;
-}
 
 export interface KnxLogger {
   get: (options?: KnxLogOptions) => Logger;
 }
 
-export interface Logger {
-  trace: (...args: any[]) => void;
-  debug: (...args: any[]) => void;
-  info: (...args: any[]) => void;
-  warn: (...args: any[]) => void;
-  error: (...args: any[]) => void;
-  format: LogDriverOptions['format'];
-}
-
 let logger: Logger;
-
 
 export interface KnxLogOptions {
   debug?: boolean;
@@ -37,7 +15,7 @@ export interface KnxLogOptions {
 
 const create = (options: KnxLogOptions): Logger => {
   const level: LogLevel =
-    (options && (options.debug ? LogLevel.Debug : options.loglevel)) || LogLevel.Info;
+    (options && (options.debug ? 'debug' : options.loglevel)) || 'info';
   return factory({
     level,
     format(lvl: LogLevel, msg: string, ...a: any[]) {
